Refactor SpamEstimator into nested String and Bike modules
